% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{build_startup_message}
\alias{build_startup_message}
\title{Build formatted startup message}
\usage{
build_startup_message(pkg_version, data_pkg_available)
}
\description{
Build formatted startup message
}
\keyword{internal}
