% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_data_required}
\alias{check_data_required}
\title{Check if data package is required and available}
\usage{
check_data_required()
}
\description{
Check if data package is required and available
}
\keyword{internal}
