% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manage.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{Check Internet Connection}
\usage{
check_internet_connection()
}
\value{
Logical. TRUE if internet is available, FALSE otherwise.
}
\description{
Helper function to check if internet connection is available
}
\keyword{internal}
