% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_update_status_silent}
\alias{check_update_status_silent}
\title{Check update status silently (for startup message)}
\usage{
check_update_status_silent()
}
\description{
Check update status silently (for startup message)
}
\keyword{internal}
