% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manage.R
\name{compare_data_versions}
\alias{compare_data_versions}
\title{Compare Data Versions}
\usage{
compare_data_versions(local_metadata, web_info)
}
\arguments{
\item{local_metadata}{List. Local metadata information.}

\item{web_info}{List. Web information from get_latest_reptile_download.}
}
\value{
List with comparison results.
}
\description{
Helper function to compare local metadata with web information
}
\keyword{internal}
