% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{extract_attribute}
\alias{extract_attribute}
\title{Extract and optionally format a specific attribute}
\usage{
extract_attribute(reptile_data, attribute_name, format_function = NULL)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}

\item{attribute_name}{Name of the attribute to extract (e.g., "Distribution", "Synonym").}

\item{format_function}{Optional formatting function for custom processing.}
}
\value{
A tibble with formatted information for the specified attribute.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Extracts and optionally formats structured information for a given attribute.
}
