% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_all_attributes}
\alias{format_all_attributes}
\title{Format all available reptile attributes}
\usage{
format_all_attributes(reptile_data, quiet = FALSE)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}

\item{quiet}{Logical. If TRUE, suppresses informational messages. Default is FALSE.}
}
\value{
A list containing all formatted attributes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Applies formatting functions to all known attributes in the reptile dataset.
}
