% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_comments}
\alias{format_comments}
\title{Format reptile comment data}
\usage{
format_comments(reptile_data)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}
}
\value{
A tibble containing formatted comments for each species.
}
\description{
Extracts and formats general comments associated with reptile species.
}
