% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_common_names}
\alias{format_common_names}
\title{Format common names for reptiles
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
format_common_names(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted common names by species.
}
\description{
Format common names for reptiles
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
