% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_diagnosis}
\alias{format_diagnosis}
\title{Format diagnostic information for reptiles}
\usage{
format_diagnosis(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted diagnostic descriptions by species.
}
\description{
Format diagnostic information for reptiles
}
