% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_reproduction}
\alias{format_reproduction}
\title{Format reproductive data for reptiles}
\usage{
format_reproduction(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted reproductive information by species.
}
\description{
Format reproductive data for reptiles
}
