% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_selected_attributes}
\alias{format_selected_attributes}
\title{Format selected reptile attributes}
\usage{
format_selected_attributes(reptile_data, attributes, quiet = FALSE)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}

\item{attributes}{A character vector specifying which attributes to extract.}

\item{quiet}{Logical. If TRUE, suppresses informational messages. Default is FALSE.}
}
\value{
A list containing the selected formatted attributes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Extracts and formats only the specified attributes from the reptile dataset.
}
