% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_synonyms}
\alias{format_synonyms}
\title{Format synonym data for reptiles}
\usage{
format_synonyms(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted synonym names by species.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
