% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manage.R
\name{print_reptile_status}
\alias{print_reptile_status}
\title{Print Status Summary}
\usage{
print_reptile_status(status)
}
\arguments{
\item{status}{List. Output from check_reptile_data_status.}
}
\value{
None. Prints formatted output.
}
\description{
Helper function to print a formatted summary of the data status
}
