% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptiledbr-package.R
\docType{package}
\name{reptiledbr-package}
\alias{reptiledbr}
\alias{reptiledbr-package}
\title{reptiledbr: Interface to the Reptile Database for Querying and Retrieving Taxonomic Data}
\description{
Provides tools to search, access, and format taxonomic information from the Reptile Database (\url{http://reptile-database.org}) directly within R. Users can retrieve species-level data, distribution, etymology, synonyms, common names, and other relevant information for reptiles. Designed for taxonomists, ecologists, and biodiversity researchers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/reptiledbr}
  \item \url{https://paulesantos.github.io/reptiledbr/}
  \item Report bugs at \url{https://github.com/PaulESantos/reptiledbr/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

}
\keyword{internal}
