% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_etymology}
\alias{format_etymology}
\title{Format etymological data of reptiles
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Extracts and formats the etymological information associated with reptile species.}
\usage{
format_etymology(reptile_data)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}
}
\value{
A tibble containing the formatted etymological details for each species.
}
\description{
Format etymological data of reptiles
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Extracts and formats the etymological information associated with reptile species.
}
