% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_higher_taxa}
\alias{format_higher_taxa}
\title{Format higher-level taxonomic data for reptiles}
\usage{
format_higher_taxa(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted higher taxonomic classification by species.
}
\description{
Format higher-level taxonomic data for reptiles
}
