% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_references}
\alias{format_references}
\title{Format bibliographic reference data of reptiles}
\usage{
format_references(reptile_data)
}
\arguments{
\item{reptile_data}{A data frame obtained using the \code{get_reptile_species_data()} function.}
}
\value{
A tibble containing formatted references for each species.
}
\description{
Extracts and formats bibliographic references associated with reptile species.
}
