% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_types}
\alias{format_types}
\title{Format nomenclatural type data for reptiles}
\usage{
format_types(reptile_data)
}
\arguments{
\item{reptile_data}{A tibble returned by \code{get_reptile_species_data()}.}
}
\value{
A tibble with formatted type information by species.
}
\description{
Format nomenclatural type data for reptiles
}
